backend/requirements/base.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
boto3==1.34.34
django-storages==1.14.2
Pillow==10.2.0

# AI & ML
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
sentence-transformers==2.3.1
google-cloud-vision==3.7.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27

# Integrations
python-telegram-bot==20.8
twilio==8.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Payment
stripe==8.1.0

# Security & Utils
cryptography==42.0.2
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0

# Validation
pydantic==2.6.0
phonenumbers==8.13.29
